Crate remotery

Remotery is a realtime CPU/GPU profiler with a viewer that runs in a web browser. This lib is a Rust wrapper around the C API provided by Remotery and the original repo over here https://github.com/Celtoys/Remotery

Modules

error

Structs

Remotery
Holds the main instance for Remotery

RemoteryScope
Scopes allows you to profile a bit of code and the end_cpu_sample will be called once it goes out of scope

Enums

SampleFlags
Sample flags to decide how profiling info should be handled