Enum remotery::error::RemoteryError

pub enum RemoteryError {
    Malloc,
    TlsAllocFail,
    VirtualMemoryBufferFail,
    CreateThreadFail,
    SocketInitNetworkFail,
    SocketCreateFail,
    SocketBindFail,
    SocketListenFail,
    SocketSetNonBlockingFail,
    SocketInvalidPoll,
    SocketSelectFail,
    SocketPollErrors,
    SocketAcceptFail,
    SocketSendTimeout,
    SocketSendFail,
    SocketRecvNoData,
    SocketRecvTimeout,
    SocketRecvFailed,
    WebsocketHandshakeNotGet,
    WebsocketHandshakeNoVersion,
    WebsocketHandshakeBadVersion,
    WebsocketHandshakeNoHost,
    WebsocketHandshakeBadHost,
    WebsocketHandshakeNoKey,
    WebsocketHandshakeBadKey,
    WebsocketHandshakeStringFail,
    WebsocketDisconnected,
    WebsocketBadFrameHeader,
    WebsocketBadFrameHeaderSize,
    WebsocketBadFrameHeaderMask,
    WebsocketReceiveTimeout,
    RemoteryNotCreated,
    SendOnIncompleteProfile,
    CudaDeinitialized,
    CudaNotInitialized,
    CudaInvalidContext,
    CudaInvalidValue,
    CudaInvalidHandle,
    CudaOutofMemory,
    ErrorNotReady,
    D3d11FailedToCreateQuery,
    OpenglError,
    Unknown,
}

Variants

Malloc

Malloc call within remotery failed

TlsAllocFail

Attempt to allocate thread local storage failed

VirtualMemoryBufferFail

Failed to create a virtual memory mirror buffer

CreateThreadFail

Failed to create a thread for the server

SocketInitNetworkFail

Network initialisation failure (e.g. on Win32, WSAStartup fails)

SocketCreateFail

Can't create a socket for connection to the remote viewer

SocketBindFail

Can't bind a socket for the server

SocketListenFail

Created server socket failed to enter a listen state

SocketSetNonBlockingFail

Created server socket failed to switch to a non-blocking state

SocketInvalidPoll

Poll attempt on an invalid socket

SocketSelectFail

Server failed to call select on socket

SocketPollErrors

Poll notified that the socket has errors

SocketAcceptFail

Server failed to accept connection from client

SocketSendTimeout

Timed out trying to send data

SocketSendFail

Unrecoverable error occured while client/server tried to send data

SocketRecvNoData

No data available when attempting a receive

SocketRecvTimeout

Timed out trying to receive data

SocketRecvFailed

Unrecoverable error occured while client/server tried to receive data

WebsocketHandshakeNotGet

WebSocket server handshake failed, not HTTP GET

WebsocketHandshakeNoVersion

WebSocket server handshake failed, can't locate WebSocket version

WebsocketHandshakeBadVersion

WebSocket server handshake failed, unsupported WebSocket version

WebsocketHandshakeNoHost

WebSocket server handshake failed, can't locate host

WebsocketHandshakeBadHost

WebSocket server handshake failed, host is not allowed to connect

WebsocketHandshakeNoKey

WebSocket server handshake failed, can't locate WebSocket key

WebsocketHandshakeBadKey

WebSocket server handshake failed, WebSocket key is ill-formed

WebsocketHandshakeStringFail

WebSocket server handshake failed, internal error, bad string code

WebsocketDisconnected

WebSocket server received a disconnect request and closed the socket

WebsocketBadFrameHeader

Couldn't parse WebSocket frame header

WebsocketBadFrameHeaderSize

Partially received wide frame header size

WebsocketBadFrameHeaderMask

Partially received frame header data mask

WebsocketReceiveTimeout

Timeout receiving frame header

RemoteryNotCreated

Remotery object has not been created

SendOnIncompleteProfile

An attempt was made to send an incomplete profile tree to the client

CudaDeinitialized

This indicates that the CUDA driver is in the process of shutting down

CudaNotInitialized

This indicates that the CUDA driver has not been initialized with cuInit() or that initialization has failed

CudaInvalidContext

This most frequently indicates that there is no context bound to the current thread

CudaInvalidValue

This indicates that one or more of the parameters passed to the API call is not within an acceptable range of values

CudaInvalidHandle

This indicates that a resource handle passed to the API call was not valid

CudaOutofMemory

The API call failed because it was unable to allocate enough memory to perform the requested operation

ErrorNotReady

This indicates that a resource handle passed to the API call was not valid

D3d11FailedToCreateQuery

Failed to create query for sample

OpenglError

Generic OpenGL error, no real need to expose more detail since app will probably have an OpenGL error callback registered

Unknown

Unknown error